# packages/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: OpenVMS API
  version: 0.3.0
  description: |
    Control plane API for the Federated Frigate VMS.
    Every endpoint is defined here first; server and client code are generated from this file.
servers:
  - url: /
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      tags: [identity]
      operationId: login
      summary: Log in with username and password (and TOTP when enabled)
      security: []
      description: |
        On success sets the HttpOnly session cookie. When the account has MFA and no
        totp_code was sent, answers mfa_required without opening a session. After
        repeated failures the account locks for a while (429).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Logged in, or a TOTP code is needed
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429":
          description: Account temporarily locked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API token for scripts and integrations. Browsers use the openvms_session cookie set by login.
    cookieAuth:
      type: apiKey
      in: cookie
      name: openvms_session
      description: Browser session. State-changing requests must also send the X-OpenVMS-Request header.
  responses:
...
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Me:
      type: object
      required: [id, username, display_name, tenant_id, grants, mfa_enabled, must_change_password, auth_method]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        display_name: { type: string }
        mfa_enabled: { type: boolean }
        must_change_password: { type: boolean }
        auth_method:
          type: string
          enum: [session, token]
          description: How this request was authenticated.
        tenant_id:
          type: string
          format: uuid
          nullable: true
          description: Null for platform users.
        grants:
          type: array
          items: { $ref: "#/components/schemas/Grant" }
    SubjectType:
      type: string
      enum: [user, group]
    ScopeType:
      type: string
      enum: [platform, tenant, site, server, camera_group, camera]
    Effect:
      type: string
      enum: [allow, deny]
...
    GrantInput:
      type: object
      required: [subject_type, subject_id, permission, effect, scope_type]
      properties:
        subject_type: { $ref: "#/components/schemas/SubjectType" }
        subject_id: { type: string, format: uuid }
        permission: { type: string, example: live.view }
        effect: { $ref: "#/components/schemas/Effect" }
        scope_type: { $ref: "#/components/schemas/ScopeType" }
        scope_id:
          type: string
          format: uuid
          description: Required unless scope_type is platform.
    Grant:
      allOf:
        - $ref: "#/components/schemas/GrantInput"
        - type: object
          required: [id, created_at]
          properties:
            id: { type: string, format: uuid }
            tenant_id: { type: string, format: uuid, nullable: true }
            created_at: { type: string, format: date-time }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 64 }
        password: { type: string, format: password, minLength: 1, maxLength: 256 }
        totp_code: { type: string, maxLength: 7 }
    LoginResponse:
      type: object
      required: [mfa_required]
      properties:
        mfa_required:
          type: boolean
          description: True when the password was right and a TOTP code must be sent too.
        expires_at: { type: string, format: date-time }
        user: { $ref: "#/components/schemas/Me" }
...